StatCard: Cover loading
